Add files via upload
This is my professional resume showcasing my expertise in web development, UI/UX design, and programming. It highlights my academic background, technical skills in Python, C++, and JavaScript, and hands-on experience with tools like Figma and Adobe XD. I am applying for aiming to contribute with innovative solutions and a strong passion for technology.